enterprise-data-visualization-app-optimized/backend/requirements.txt
```
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9 # PostgreSQL adapter
Flask-RESTX==1.2.0 # API development and Swagger UI
Flask-JWT-Extended==4.4.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
python-decouple==3.8 # For env vars

# Additional Features
Flask-Caching==2.0.2
Flask-Limiter==3.5.1
redis==5.0.1 # For caching and rate limiting with Redis
SQLAlchemy-JSONField==0.2.0 # For storing flexible JSON data in models

# Data source connectors (examples)
pandas==2.1.3
sqlalchemy-access==1.1.0 # Example for MS Access, can be any DB specific SQLAlchemy dialect
PyMySQL==1.1.0 # Example for MySQL
psycopg2-binary # Already listed, but for explicit clarity

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
Faker==20.0.0 # For test data generation
locust==2.19.0 # Performance testing
```